% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Yosemite.R
\docType{data}
\name{Yosemite}
\alias{Yosemite}
\title{Spatial data for western fence lizards in Yosemite}
\format{A RasterStack}
\usage{
data(Yosemite)
}
\description{
A dataset containing a RasterStack with two RasterLayer objects, a digital elevation model (DEM)
for part of Yosemite National Park (USA) and a species distribution model (SDM) for the western
fence lizard (Sceloporus occidentalis).
}
\examples{
data(Yosemite)
}
\keyword{datasets}
